+++ Lib/TSCompiler/tslib.js
@@ -143,3 +143,8 @@ var __asyncValues = function (o) {
     var m = o[Symbol.asyncIterator];
     return m ? m.call(o) : typeof __values === "function" ? __values(o) : o[Symbol.iterator]();
 };
+
+var __makeTemplateObject = function (cooked, raw) {
+    Object.defineProperty(cooked, "raw", { value: raw });
+    return cooked;
+};